+++ packages/provider-elasticsearch/src/example-types/metricGroups/pivot.js
@@ -67,7 +67,7 @@ let maybeWrapWithFilterAgg = ({
         },
       }
 
-let getKey = (x) => x.keyAsString || x.key
+let getKey = x => x.keyAsString || x.key
 
 // Similar to Tree.lookup but path is a drilldown which uses keyAsString or key
 let resultsForDrilldown = (groupType, drilldown, results) => {
@@ -75,7 +75,7 @@ let resultsForDrilldown = (groupType, drilldown, results) => {
 
   let key = _.first(drilldown)
   let groups = _.get(groupType, results)
-  let match = _.find((node) => getKey(node) === key, groups)
+  let match = _.find(node => getKey(node) === key, groups)
   return resultsForDrilldown(groupType, drilldown.slice(1), match)
 }
 
@@ -88,6 +88,15 @@ let mergeResults = _.mergeWith((current, additional, prop) => {
     )([current, additional])
   } else if (_.isArray(additional)) return additional
 })
+
+//Functions for hoisting functionality
+let hoistProps = {}
+let removeHoistProps = _.curry(group => _.omit(['hoistProps'], group))
+let mergeHoistProps = _.curry((props, group) =>
+  _.merge(_.getOr({}, 'hoistProps', group), props)
+)
+let mergeNestedHoistProps = _.curry(_.reduce(mergeHoistProps(hoistProps), {}))
+let removeNestedHoistProps = _.curry(_.map(removeHoistProps))
 
 let createPivotScope = (node, schema, getStats) => {
   /***
@@ -126,7 +135,7 @@ let createPivotScope = (node, schema, getStats) => {
    ***/
   let findNotLoadedExpansion = () => _.find(({ loaded }) => !loaded, expansions)
 
-  let previouslyLoadedKeys = (expansion) =>
+  let previouslyLoadedKeys = expansion =>
     _.flow(
       _.filter(
         ({ type, drilldown, loaded }) =>
@@ -137,10 +146,10 @@ let createPivotScope = (node, schema, getStats) => {
       _.flatMap('loaded')
     )(expansions)
 
-  let isFirstOfType = (expansion) =>
+  let isFirstOfType = expansion =>
     expansion === _.find({ type: expansion.type }, expansions)
 
-  let getGridExpansions = (expansion) =>
+  let getGridExpansions = expansion =>
     _.flow(
       _.take(_.indexOf(expansion, expansions)),
       _.filter(({ type }) => type !== expansion.type)
@@ -151,7 +160,7 @@ let createPivotScope = (node, schema, getStats) => {
   }
 
   // Merge current row/column drilldown with includes from already expanded columns/rows
-  let makeRequest = (expansion, values) => (gridExpansion) => {
+  let makeRequest = (expansion, values) => gridExpansion => {
     let rowDrillMode = expansion.type === 'rows'
     return {
       type: expansion.type,
@@ -170,7 +179,7 @@ let createPivotScope = (node, schema, getStats) => {
     }
   }
 
-  let getInitialRequest = (expansion) =>
+  let getInitialRequest = expansion =>
     makeRequest(expansion)(_.first(getGridExpansions(expansion)))
 
   let getAdditionalRequests = (expansion, includeValues) =>
@@ -184,7 +193,7 @@ let createPivotScope = (node, schema, getStats) => {
    ***/
 
   // Builds aggregation for pivot values
-  let getAggsForValues = (values) =>
+  let getAggsForValues = values =>
     _.flow(
       _.keyBy(aggKeyForValue),
       // This is a very interesting lint error - while key is not used in the function body, it is important.
@@ -221,14 +230,14 @@ let createPivotScope = (node, schema, getStats) => {
     let filter = lookupTypeProp(_.stubFalse, 'drilldown', group.type)
     return _.flatten(
       await compactMapAsync(
-        (value) => filter({ drilldown: value, ...group }, schema, getStats),
+        value => filter({ drilldown: value, ...group }, schema, getStats),
         values
       )
     )
   }
 
   // Builds aggregation for sorting
-  let getSortAgg = async (sort) => {
+  let getSortAgg = async sort => {
     if (!sort) return
     let filters = await getDrilldownFilters({
       drilldown: sort.columnValues,
@@ -321,17 +330,39 @@ let createPivotScope = (node, schema, getStats) => {
     )
   }
 
-  let buildQuery = async (request) => {
+  let buildQuery = async request => {
     let columnDrills = _.getOr([], 'columns.drilldown', request)
     let rowDrills = _.getOr([], 'rows.drilldown', request)
+    hoistProps = {}
     // Don't consider deeper levels than +1 the current drilldown
     // This allows avoiding expansion until ready
     // Opt out with expandColumns / expandRows
+    let query
 
-    let hoistProps = {}
-    let mergeHoistProps = (hoistProps, statsAggs) =>
-      _.merge(_.getOr({}, 'hoistProps', statsAggs), hoistProps)
-    let removeHoistProps = (statsAggs) => _.omit('hoistProps', statsAggs)
+    let axisFilters = async ({ drills, axisGroup, axisName, hoistProps }) => {
+      // Filtering data specified by the drilldown
+      let drillDownFilters = await getDrilldownFilters({
+        drilldown: drills,
+        groups: axisGroup,
+      })
+      // Narrowing query down to values specified in include
+      let includeFilters = await getRequestFilters({
+        drilldown: drills,
+        values: _.get(`${axisName}.include`, request),
+        groups: axisGroup,
+      })
+      // Skipping existing values when requesting the next request
+      let skipFilters = await getRequestFilters({
+        drilldown: drills,
+        values: _.get(`${axisName}.skip`, request),
+        groups: axisGroup,
+      })
+      hoistProps = mergeNestedHoistProps(drillDownFilters)
+      hoistProps = mergeNestedHoistProps(includeFilters)
+      drillDownFilters = removeNestedHoistProps(drillDownFilters)
+      includeFilters = removeNestedHoistProps(includeFilters)
+      return [drillDownFilters, includeFilters, skipFilters, hoistProps]
+    }
 
     let columns = _.get('expanded.columns', node)
       ? node.columns
@@ -341,41 +372,20 @@ let createPivotScope = (node, schema, getStats) => {
       ? node.rows
       : _.take(_.size(rowDrills) + 1, node.rows)
 
-    // Filtering data specified by the drilldown
-    let drilldownColumnFilters = await getDrilldownFilters({
-      drilldown: columnDrills,
-      groups: columns,
-    })
-    let drilldownRowFilters = await getDrilldownFilters({
-      drilldown: rowDrills,
-      groups: rows,
-    })
-
-    // Narrowing query down to values specified in include
-    let includeColumnFilters = await getRequestFilters({
-      drilldown: columnDrills,
-      values: _.get('columns.include', request),
-      groups: columns,
-    })
-    let includeRowFilters = await getRequestFilters({
-      drilldown: rowDrills,
-      values: _.get('rows.include', request),
-      groups: rows,
-    })
-
-    // Skipping existing values when requesting the next request
-    let skipColumnFilters = await getRequestFilters({
-      drilldown: columnDrills,
-      values: _.get('columns.skip', request),
-      groups: columns,
-    })
-    let skipRowFilters = await getRequestFilters({
-      drilldown: rowDrills,
-      values: _.get('rows.skip', request),
-      groups: rows,
-    })
-
     let statsAggs = { aggs: getAggsForValues(node.values) }
+
+    let [
+      drilldownColumnFilters,
+      includeColumnFilters,
+      skipColumnFilters,
+      colHoistProps,
+    ] = await axisFilters({
+      drills: columnDrills,
+      axisGroup: columns,
+      axisName: 'columns',
+      hoistProps,
+    })
+    hoistProps = _.merge(hoistProps, colHoistProps)
 
     if (!_.isEmpty(columns)) {
       let columnsStatsAggs = await buildNestedGroupQuery(
@@ -416,10 +426,22 @@ let createPivotScope = (node, schema, getStats) => {
       'rows',
       node.sort
     )
-    let query
 
     hoistProps = mergeHoistProps(hoistProps, rowsStatsAggs)
     rowsStatsAggs = removeHoistProps(rowsStatsAggs)
+
+    let [
+      drilldownRowFilters,
+      includeRowFilters,
+      skipRowFilters,
+      rowHoistProps,
+    ] = await axisFilters({
+      drills: rowDrills,
+      axisGroup: rows,
+      axisName: 'rows',
+      hoistProps,
+    })
+    hoistProps = _.merge(hoistProps, rowHoistProps)
 
     if (request.rows.totals) {
       // adding total rows statsAggs above the rows filters
@@ -531,7 +553,7 @@ let createPivotScope = (node, schema, getStats) => {
 let hasValue = ({ filters }) => !_.isEmpty(filters)
 let filter = async (node, schema) => {
   // This requires getting `search` passed in to filter
-  // This is a change to contexture server, which is likely a breaking change moving all contexture type methods to named object params
+  // This is a change to contexture core, which is likely a breaking change moving all contexture type methods to named object params
   //    That will allow everything to get all props inclding `search`
   let { filters, rows = [], columns = [] } = node
   let getStats = () => {
@@ -539,29 +561,46 @@ let filter = async (node, schema) => {
   }
   let { getDrilldownFilters } = createPivotScope(node, schema, getStats)
 
-  return or(
-    await compactMapAsync(
-      async (filter) =>
+  hoistProps = {}
+
+  let boolFilter =  or(
+   await compactMapAsync(
+    async filter =>
         and([
-          await getDrilldownFilters({
-            drilldown: filter.rows,
-            groups: rows,
-          }),
-          await getDrilldownFilters({
-            drilldown: filter.columns,
-            groups: columns,
-          }),
+          _.map(
+            (filterGroup) => {
+              hoistProps = mergeHoistProps(hoistProps, filterGroup)
+              return removeHoistProps(filterGroup)
+            },
+            await getDrilldownFilters({
+              drilldown: filter.rows,
+              groups: rows,
+            })
+          ),
+          _.map(filterGroup => {
+              hoistProps = mergeHoistProps(hoistProps, filterGroup)
+              return removeHoistProps(filterGroup)
+            },
+            await getDrilldownFilters({
+              drilldown: filter.columns,
+              groups: columns,
+            }),
+          ),
         ]),
-      filters
-    )
-  )
+    filters
+  ))
+
+  return {
+    hoistProps,
+    ...boolFilter,
+  }
 }
 
 export default {
   hasValue,
   filter,
   createPivotScope,
-  validContext: (node) => node.rows.length && node.values.length,
+  validContext: node => node.rows.length && node.values.length,
   async result(node, search, schema) {
     let {
       findNotLoadedExpansion,